Fix bugs in error return from writevv.
We were not properly scrubbing the td->td_retval between calls to
writevv_kernel(), this would result in the wrong errno and return
values being set (masking blocked writes).

Additionally we were not returning an indicator of full success or
not.  Now we return the number of full complete writes so that
user code can check the return value against the number of fds
to avoid checking the entire array for errors/partial writes.